feat: organizations get soft-deleted when deleted via Admin UI (#1432)

+++ server/repository/organization-repository.go
@@ -27,6 +27,8 @@ type Organization struct {
 	ContactEmail     string
 	Language         string
 	SignupDate       time.Time
+	Deleted          bool
+	DeletedAtUTC     *time.Time
 }
 
 type Domain struct {
@@ -140,15 +142,22 @@ func (r *OrganizationRepository) RunSchemaUpgrade(curVersion, targetVersion int)
 			panic(err)
 		}
 	}
+	if curVersion < 29 {
+		if _, err := GetDatabase().DB().Exec("ALTER TABLE organizations " +
+			"ADD COLUMN IF NOT EXISTS deleted boolean NOT NULL DEFAULT FALSE, " +
+			"ADD COLUMN IF NOT EXISTS deleted_at_utc TIMESTAMP NULL DEFAULT NULL"); err != nil {
+			panic(err)
+		}
+	}
 }
 
 func (r *OrganizationRepository) Create(e *Organization) error {
 	var id string
 	err := GetDatabase().DB().QueryRow("INSERT INTO organizations "+
-		"(name, contact_firstname, contact_lastname, contact_email, language, signup_date) "+
-		"VALUES ($1, $2, $3, $4, $5, $6) "+
+		"(name, contact_firstname, contact_lastname, contact_email, language, signup_date, deleted, deleted_at_utc) "+
+		"VALUES ($1, $2, $3, $4, $5, $6, $7, $8) "+
 		"RETURNING id",
-		e.Name, e.ContactFirstname, e.ContactLastname, e.ContactEmail, e.Language, e.SignupDate).Scan(&id)
+		e.Name, e.ContactFirstname, e.ContactLastname, e.ContactEmail, e.Language, e.SignupDate, e.Deleted, e.DeletedAtUTC).Scan(&id)
 	if err != nil {
 		return err
 	}
@@ -162,11 +171,27 @@ func (r *OrganizationRepository) Create(e *Organization) error {
 
 func (r *OrganizationRepository) GetOneByDomain(domain string) (*Organization, error) {
 	e := &Organization{}
-	err := GetDatabase().DB().QueryRow("SELECT organizations.id, organizations.name, organizations.contact_firstname, organizations.contact_lastname, organizations.contact_email, organizations.language, organizations.signup_date "+
+	err := GetDatabase().DB().QueryRow("SELECT organizations.id, organizations.name, organizations.contact_firstname, organizations.contact_lastname, organizations.contact_email, organizations.language, organizations.signup_date, organizations.deleted, organizations.deleted_at_utc "+
 		"FROM organizations_domains "+
 		"INNER JOIN organizations ON organizations.id = organizations_domains.organization_id "+
-		"WHERE LOWER(organizations_domains.domain) = $1 AND organizations_domains.active = TRUE",
-		strings.ToLower(domain)).Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate)
+		"WHERE LOWER(organizations_domains.domain) = $1 AND organizations_domains.active = TRUE AND organizations.deleted = FALSE",
+		strings.ToLower(domain)).Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate, &e.Deleted, &e.DeletedAtUTC)
+	if err != nil {
+		return nil, err
+	}
+	return e, nil
+}
+
+func (r *OrganizationRepository) GetOneIncludeDeleted(id string, includeDeleted bool) (*Organization, error) {
+	appendix := " AND deleted = FALSE"
+	if includeDeleted {
+		appendix = ""
+	}
+	e := &Organization{}
+	err := GetDatabase().DB().QueryRow("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date, deleted, deleted_at_utc "+
+		"FROM organizations "+
+		"WHERE id = $1"+appendix,
+		id).Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate, &e.Deleted, &e.DeletedAtUTC)
 	if err != nil {
 		return nil, err
 	}
@@ -174,23 +199,15 @@ func (r *OrganizationRepository) GetOneByDomain(domain string) (*Organization, e
 }
 
 func (r *OrganizationRepository) GetOne(id string) (*Organization, error) {
-	e := &Organization{}
-	err := GetDatabase().DB().QueryRow("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date "+
-		"FROM organizations "+
-		"WHERE id = $1",
-		id).Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate)
-	if err != nil {
-		return nil, err
-	}
-	return e, nil
+	return r.GetOneIncludeDeleted(id, false)
 }
 
 func (r *OrganizationRepository) GetByEmail(email string) (*Organization, error) {
 	e := &Organization{}
-	err := GetDatabase().DB().QueryRow("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date "+
+	err := GetDatabase().DB().QueryRow("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date, deleted, deleted_at_utc "+
 		"FROM organizations "+
 		"WHERE LOWER(contact_email) = $1",
-		strings.ToLower(email)).Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate)
+		strings.ToLower(email)).Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate, &e.Deleted, &e.DeletedAtUTC)
 	if err != nil {
 		return nil, err
 	}
@@ -199,15 +216,15 @@ func (r *OrganizationRepository) GetByEmail(email string) (*Organization, error)
 
 func (r *OrganizationRepository) GetAll() ([]*Organization, error) {
 	var result []*Organization
-	rows, err := GetDatabase().DB().Query("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date " +
-		"FROM organizations ORDER BY name")
+	rows, err := GetDatabase().DB().Query("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date, deleted, deleted_at_utc " +
+		"FROM organizations WHERE deleted = FALSE ORDER BY name")
 	if err != nil {
 		return nil, err
 	}
 	defer rows.Close()
 	for rows.Next() {
 		e := &Organization{}
-		err = rows.Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate)
+		err = rows.Scan(&e.ID, &e.Name, &e.ContactFirstname, &e.ContactLastname, &e.ContactEmail, &e.Language, &e.SignupDate, &e.Deleted, &e.DeletedAtUTC)
 		if err != nil {
 			return nil, err
 		}
@@ -225,7 +242,7 @@ func (r *OrganizationRepository) GetAllDaysPassedSinceSignup(daysPassed int, set
 	}
 	rows, err := GetDatabase().DB().Query("SELECT id, name, contact_firstname, contact_lastname, contact_email, language, signup_date "+
 		"FROM organizations "+
-		"WHERE (CURRENT_DATE::date - signup_date::date) = $1 "+
+		"WHERE deleted = FALSE AND (CURRENT_DATE::date - signup_date::date) = $1 "+
 		settingExistsQuery+" "+
 		"ORDER BY name", interval, settingExists)
 	if err != nil {
@@ -245,7 +262,7 @@ func (r *OrganizationRepository) GetAllDaysPassedSinceSignup(daysPassed int, set
 
 func (r *OrganizationRepository) GetNumOrgs() (int, error) {
 	var result int
-	err := GetDatabase().DB().QueryRow("SELECT COUNT(*) FROM organizations").Scan(&result)
+	err := GetDatabase().DB().QueryRow("SELECT COUNT(*) FROM organizations WHERE deleted = FALSE").Scan(&result)
 	if err != nil {
 		return 0, err
 	}
@@ -255,7 +272,7 @@ func (r *OrganizationRepository) GetNumOrgs() (int, error) {
 func (r *OrganizationRepository) GetAllIDs() ([]string, error) {
 	var result []string
 	rows, err := GetDatabase().DB().Query("SELECT id " +
-		"FROM organizations")
+		"FROM organizations WHERE deleted = FALSE")
 	if err != nil {
 		return nil, err
 	}
@@ -285,9 +302,42 @@ func (r *OrganizationRepository) Update(e *Organization) error {
 	return nil
 }
 
-func (r *OrganizationRepository) Delete(e *Organization) error {
+func (r *OrganizationRepository) RestoreSoftDeleted(e *Organization) error {
+	_, err := GetDatabase().DB().Exec("UPDATE organizations SET "+
+		"deleted = FALSE, deleted_at_utc = NULL "+
+		"WHERE id = $1 AND deleted = TRUE",
+		e.ID)
+	if err != nil {
+		return err
+	}
+	e.Deleted = false
+	e.DeletedAtUTC = nil
+	return nil
+}
+
+func (r *OrganizationRepository) DeleteSoft(e *Organization) error {
 	for _, plg := range plugin.GetPlugins() {
-		(*plg).OnBeforeOrganizationDelete(e.ID)
+		(*plg).OnBeforeOrganizationDeleteSoft(e.ID)
+	}
+	now := time.Now().UTC()
+	_, err := GetDatabase().DB().Exec("UPDATE organizations SET "+
+		"deleted = TRUE, deleted_at_utc = $1 "+
+		"WHERE id = $2",
+		now, e.ID)
+	if err != nil {
+		return err
+	}
+	e.Deleted = true
+	e.DeletedAtUTC = &now
+	return nil
+}
+
+func (r *OrganizationRepository) DeleteHard(e *Organization) error {
+	if !e.Deleted {
+		return errors.New("organization must be soft-deleted before hard deletion")
+	}
+	for _, plg := range plugin.GetPlugins() {
+		(*plg).OnBeforeOrganizationDeleteHard(e.ID)
 	}
 	if err := GetAuthProviderRepository().DeleteAll(e.ID); err != nil {
 		return err
